+++ Tests/FluentAssertions.Specs/AndWhichConstraintSpecs.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using FluentAssertions.Collections;
+using FluentAssertions.Execution;
 using Xunit;
 using Xunit.Sdk;
 
@@ -11,7 +12,7 @@ public class AndWhichConstraintSpecs
     public void When_many_objects_are_provided_accessing_which_should_throw_a_descriptive_exception()
     {
         // Arrange
-        var continuation = new AndWhichConstraint<StringCollectionAssertions, string>(null, ["hello", "world"]);
+        var continuation = new AndWhichConstraint<StringCollectionAssertions, string>(null, ["hello", "world"], AssertionChain.GetOrCreate());
 
         // Act
         Action act = () => _ = continuation.Which;